Make ORKCustomStepView.h public and ensure it's not stripped in Release.
RK's podspec exports all private headers as public, so this header was
public when we created our custom step for optional dates.

Now that we use Carthage, it relies on the xcodeproj file instead, and
the header being private causes linker errors when building for Release.
This appears to be the only real option for customizing steps, so we
are modifying our fork and will submit an issue/PR with RK upstream.